+++ frontend/web/src/features/hostdashboard/website/WebsiteBuilderPage.js
@@ -43,6 +43,7 @@ import { applyWebsiteDraftContentOverrides } from "./rendering/websiteDraftConte
 import { applyWebsiteDraftThemeOverrides } from "./rendering/websiteDraftThemeOverrides";
 import { placeholderImage, resolveAccommodationImageUrl } from "../../../utils/accommodationImage";
 import { WEBSITE_DRAFT_DELETE_REASONS } from "./websiteDeleteReasons";
+import { buildPublishedWebsiteHref } from "./websitePublicSiteLinks";
 
 const EMPTY_SELECTION = "";
 const PHOTO_CARD_VARIANT_CLASSES = [styles.photoCard1, styles.photoCard2, styles.photoCard3];
@@ -134,24 +135,6 @@ const getDraftDisplayTitle = (draft, contentOverrides = getDraftContentOverrides
     String(draft?.propertyTitle || "").trim() ||
     "Untitled listing website"
   );
-};
-
-const buildPublishedWebsitePath = (domain, siteId = "") => {
-  const path = `/website-live/${encodeURIComponent(domain)}`;
-  const normalizedSiteId = String(siteId || "").trim();
-  return normalizedSiteId ? `${path}?siteId=${encodeURIComponent(normalizedSiteId)}` : path;
-};
-const buildPublishedWebsiteHref = (domain, siteId = "", domainStatus = "") => {
-  const normalizedDomain = String(domain || "").trim().toLowerCase();
-  if (!normalizedDomain) {
-    return "";
-  }
-
-  if (String(domainStatus || "").trim().toUpperCase() === "ACTIVE") {
-    return `https://${normalizedDomain}`;
-  }
-
-  return buildPublishedWebsitePath(normalizedDomain, siteId);
 };
 
 const buildWebsitePreviewPath = (draftId) => `/website-preview/${encodeURIComponent(draftId)}`;
